+++ src/main/java/br/com/api/service/ServicoVacina.java
@@ -1,0 +1,54 @@
+package br.com.api.service;
+
+import java.util.ArrayList;
+
+import br.com.api.model.Vacinas.Vacinas;
+import br.com.api.model.Vacinas.Vacinas.PublicoAlvo;
+
+
+public class ServicoVacina {
+    private ArrayList<Vacinas> vacinas = new ArrayList<>();
+    private int contadorId = 1;
+
+    public ArrayList<Vacinas> consultarVacinas() {
+        return vacinas;
+    }
+
+
+    public Vacinas adicionar(String vacina, int limiteAplicacao, PublicoAlvo publicoAlvo ) {
+        Vacinas usuario = new Vacinas(String.valueOf(contadorId++), vacina, limiteAplicacao, publicoAlvo );
+        vacinas.add(usuario);
+        return usuario;
+    }
+
+    public Vacinas consultarPorFaixa(PublicoAlvo publicoAlvo){
+        for(Vacinas vacina : vacinas){
+            if (vacina.getPublicoAlvo().equals(publicoAlvo) ) {
+                return vacina;
+            
+        }
+    }
+    return null;
+    }
+
+    public Vacinas consultarVacinasRecomendadas(int limiteAplicacao){
+        for(Vacinas vacina :vacinas){
+            if (vacina.getlimiteAplicacao() == (limiteAplicacao)) {
+                return vacina;
+            }
+        }
+        return null;
+    }
+    
+    public Vacinas consultarVacinasNaoAplicaveis(String id){
+        for (Vacinas vacina : vacinas) {
+            if (vacina.getId().equals(id)) {
+                return vacina;
+            }
+        }
+        return null;
+
+    }
+
+
+}
